fix: error in ComplexMath library decreased accuracy add: some official interfaces to integration library add: fit functions add: StableDistributions are much more advanced now
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%40573

+++ Altaxo/Core/Calc/complex.cs
@@ -493,8 +493,8 @@ namespace Altaxo.Calc
     /// <returns></returns>
     public static Complex operator-( double f, Complex a ) 
     {
-      a.Re  = (float)( f - a.Re );
-      a.Im  = (float)( 0 - a.Im );
+      a.Re  = ( f - a.Re );
+      a.Im  = ( 0 - a.Im );
       return a;
     }
 
@@ -519,8 +519,8 @@ namespace Altaxo.Calc
     /// <returns></returns>
     public static Complex operator*( Complex a, double f ) 
     {
-      a.Re  = (double)( a.Re * f );
-      a.Im  = (double)( a.Im * f );
+      a.Re  = ( a.Re * f );
+      a.Im  = ( a.Im * f );
       return a;
     }
